chore(deps): Update and reconfigure needed packages (#46)
Upgraded the package-lock version. NPM 6 was the last version to require a v1 lock file.

Upgraded the semantic release packages. Moved them to `peerDependencies` as these packages are both exposed to, and expected to used by, the consuming code.

Upgraded Snyk following the Snyk protect upgrade guide. Moved to being a dev dependency as Snyk isn't needed by the consumer of this package.

Co-authored-by: Ricky C <ricky.curtice@onerail.io>